chore: apply the release audit (#102)
`make audit` before cutting a release. Kept to what it flags and the project
does not deliberately allow:

- sorted `__all__` in three modules
- `itertools.pairwise` where a `zip(x, x[1:])` was walking successive pairs
- dropped six `int()` casts on values that were already integers
- collapsed two nested `if`s
- en dashes and multiplication signs out of the docstrings, hyphen and x instead

Left alone, on purpose: S603 (every ffmpeg call is a subprocess), S311 (seeded
random for B-roll), ARG002 (the block interface), PERF203 (try/except in a loop
is what per-step error handling is), and the RUF100 `# noqa`s, which document
exactly what this audit reports.

`_check_block_params` remains at complexity 48, four times the threshold. It is
a per-block if-chain; splitting it is a real refactor, not an audit fix.

+++ src/lemontage/engine/assformat.py
@@ -115,7 +115,7 @@ def timestamp(seconds: float) -> str:
     naive formatting of ``-0.5`` is ``-1:59:59.50`` — a time libass reads as
     valid but far past the clip, so the line silently never shows.
     """
-    cs = int(round(max(0.0, seconds) * 100))
+    cs = round(max(0.0, seconds) * 100)
     hours, cs = divmod(cs, 360000)
     minutes, cs = divmod(cs, 6000)
     secs, cs = divmod(cs, 100)
